refactor: update Storybook integration and remove unused components
